lfork fix

+++ vm/srcs/op_func/op_lfork.c
@@ -6,7 +6,7 @@
 /*   By: mgonon <mgonon@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/03/07 23:54:41 by gudemare          #+#    #+#             */
-/*   Updated: 2018/03/14 19:03:19 by mgonon           ###   ########.fr       */
+/*   Updated: 2018/03/14 20:17:17 by mgonon           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,10 @@
 
 int	do_lfork(t_process *process, t_env *env)
 {
-	short			param;
+	unsigned int	param;
 	t_process		*fork;
 
-	param = (short)get_param_raw_value(env, process->pc + 1, T_DIR, OP_FORK);
+	param = get_param_raw_value(env, process->pc + 1, T_DIR, OP_FORK);
 	add_new_process(env, process->champ_id);
 	fork = (t_process *)env->process->content;
 	memcpy(fork, process, sizeof(t_process));
